Brought out render command out of main
