+++ bot/commands/__init__.py
@@ -2,6 +2,7 @@ from .request import command_request
 from .avaliable import command_avaliable
 from .qa import command_qa
 from .start import command_start
+from ._render_menu import render_menu
 
 __ALL__ = ['command_request', 'command_avaliable','qa',
-           'command_start']
+           'command_start', 'render_menu']
